Refactor construction of index queries
