+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/query/graph/GraphCentricQueryBuilder.java
@@ -16,12 +16,11 @@ package org.janusgraph.graphdb.query.graph;
 
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Iterables;
-import org.apache.commons.lang3.mutable.MutableBoolean;
 import org.janusgraph.core.*;
 import org.janusgraph.core.attribute.Cmp;
 import org.janusgraph.core.attribute.Contain;
-import org.janusgraph.core.schema.SchemaStatus;
 import org.janusgraph.graphdb.database.IndexSerializer;
+import org.janusgraph.graphdb.query.index.IndexSelectionStrategy;
 import org.janusgraph.graphdb.internal.ElementCategory;
 import org.janusgraph.graphdb.internal.InternalRelationType;
 import org.janusgraph.graphdb.internal.Order;
@@ -31,14 +30,10 @@ import org.janusgraph.graphdb.query.condition.*;
 import org.janusgraph.graphdb.query.profile.QueryProfiler;
 import org.janusgraph.graphdb.transaction.StandardJanusGraphTx;
 import org.janusgraph.graphdb.types.*;
-import org.janusgraph.graphdb.util.IndexCandidate;
-import org.janusgraph.graphdb.util.IndexCandidateGroup;
-import org.janusgraph.util.datastructures.PowerSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.*;
-import java.util.stream.StreamSupport;
 
 /**
  * Builds a {@link JanusGraphQuery}, optimizes the query and compiles the result into a {@link GraphCentricQuery} which
@@ -82,13 +77,18 @@ public class GraphCentricQueryBuilder implements JanusGraphQuery<GraphCentricQue
      * Whether smart limit adjustment is enabled
      */
     private boolean useSmartLimit;
+    /**
+     * Selection service for the best combination of indexes to be queried
+     */
+    private IndexSelectionStrategy indexSelector;
 
-    public GraphCentricQueryBuilder(StandardJanusGraphTx tx, IndexSerializer serializer) {
+    public GraphCentricQueryBuilder(StandardJanusGraphTx tx, IndexSerializer serializer, IndexSelectionStrategy indexSelector) {
         Preconditions.checkNotNull(tx);
         Preconditions.checkNotNull(serializer);
         useSmartLimit = tx.getGraph().getConfiguration().adjustQueryLimit();
         this.tx = tx;
         this.serializer = serializer;
+        this.indexSelector = indexSelector;
     }
 
     public void disableSmartLimit() {
@@ -246,7 +246,6 @@ public class GraphCentricQueryBuilder implements JanusGraphQuery<GraphCentricQue
             }
         }
 
-
         //Prepare orders
         orders.makeImmutable();
         if (orders.isEmpty()) orders = OrderList.NO_ORDER;
@@ -264,8 +263,7 @@ public class GraphCentricQueryBuilder implements JanusGraphQuery<GraphCentricQue
         });
 
         final Set<Condition> coveredClauses = new HashSet<>();
-        final MutableBoolean isSorted = new MutableBoolean(orders.isEmpty());
-        final JointIndexQuery jointQuery = selectIndices(indexCandidates, conditions, coveredClauses, isSorted);
+        final IndexSelectionStrategy.SelectedIndexQuery selectedIndex = indexSelector.selectIndices(indexCandidates, conditions, coveredClauses, orders, serializer);
 
         BackendQueryHolder<JointIndexQuery> query;
         if (!coveredClauses.isEmpty()) {
@@ -275,254 +273,11 @@ public class GraphCentricQueryBuilder implements JanusGraphQuery<GraphCentricQue
             }
             indexLimit = Math.min(HARD_MAX_LIMIT,
                 QueryUtil.adjustLimitForTxModifications(tx, coveredClauses.size(), indexLimit));
-            query = new BackendQueryHolder<>(jointQuery.updateLimit(indexLimit),
-                    coveredClauses.size() == conditions.numChildren(), isSorted.getValue());
+            query = new BackendQueryHolder<>(selectedIndex.getQuery().updateLimit(indexLimit),
+                    coveredClauses.size() == conditions.numChildren(), selectedIndex.isSorted());
         } else {
-            query = new BackendQueryHolder<>(new JointIndexQuery(), false, isSorted.getValue());
+            query = new BackendQueryHolder<>(new JointIndexQuery(), false, selectedIndex.isSorted());
         }
         return new GraphCentricQuery(resultType, conditions, orders, query, limit);
-    }
-
-    public static boolean indexCoversOrder(MixedIndexType index, OrderList orders) {
-        for (int i = 0; i < orders.size(); i++) {
-            if (!index.indexesKey(orders.getKey(i))) return false;
-        }
-        return true;
-    }
-
-    public static List<Object[]> indexCover(final CompositeIndexType index, Condition<JanusGraphElement> condition,
-                                            Set<Condition> covered) {
-        if (!QueryUtil.isQueryNormalForm(condition)) {
-            return null;
-        }
-        assert condition instanceof And;
-        if (index.getStatus()!= SchemaStatus.ENABLED) return null;
-        final IndexField[] fields = index.getFieldKeys();
-        final Object[] indexValues = new Object[fields.length];
-        final Set<Condition> coveredClauses = new HashSet<>(fields.length);
-        final List<Object[]> indexCovers = new ArrayList<>(4);
-
-        constructIndexCover(indexValues,0,fields,condition,indexCovers,coveredClauses);
-        if (!indexCovers.isEmpty()) {
-            covered.addAll(coveredClauses);
-            return indexCovers;
-        } else return null;
-    }
-
-    private static void constructIndexCover(Object[] indexValues, int position, IndexField[] fields,
-                                            Condition<JanusGraphElement> condition,
-                                            List<Object[]> indexCovers, Set<Condition> coveredClauses) {
-        if (position>=fields.length) {
-            indexCovers.add(indexValues);
-        } else {
-            final IndexField field = fields[position];
-            final Map.Entry<Condition,Collection<Object>> equalCon = getEqualityConditionValues(condition,field.getFieldKey());
-            if (equalCon!=null) {
-                coveredClauses.add(equalCon.getKey());
-                assert equalCon.getValue().size()>0;
-                for (final Object value : equalCon.getValue()) {
-                    final Object[] newValues = Arrays.copyOf(indexValues,fields.length);
-                    newValues[position]=value;
-                    constructIndexCover(newValues,position+1,fields,condition,indexCovers,coveredClauses);
-                }
-            }
-        }
-    }
-
-    public static Map.Entry<Condition,Collection<Object>> getEqualityConditionValues(
-            Condition<JanusGraphElement> condition, RelationType type) {
-        for (final Condition c : condition.getChildren()) {
-            if (c instanceof Or) {
-                final Map.Entry<RelationType,Collection> orEqual = QueryUtil.extractOrCondition((Or)c);
-                if (orEqual!=null && orEqual.getKey().equals(type) && !orEqual.getValue().isEmpty()) {
-                    return new AbstractMap.SimpleImmutableEntry(c,orEqual.getValue());
-                }
-            } else if (c instanceof PredicateCondition) {
-                final PredicateCondition<RelationType, JanusGraphRelation> atom = (PredicateCondition)c;
-                if (atom.getKey().equals(type) && atom.getPredicate()==Cmp.EQUAL && atom.getValue()!=null) {
-                    return new AbstractMap.SimpleImmutableEntry(c,Collections.singletonList(atom.getValue()));
-                }
-            }
-
-        }
-        return null;
-    }
-
-    public static Condition<JanusGraphElement> indexCover(final MixedIndexType index,
-                                                          Condition<JanusGraphElement> condition,
-                                                          final IndexSerializer indexInfo,
-                                                          final Set<Condition> covered) {
-        if (!indexInfo.features(index).supportNotQueryNormalForm() && !QueryUtil.isQueryNormalForm(condition)) {
-            return null;
-        }
-        if (condition instanceof Or) {
-            for (final Condition<JanusGraphElement> subClause : condition.getChildren()) {
-                if (subClause instanceof And) {
-                    for (final Condition<JanusGraphElement> subsubClause : condition.getChildren()) {
-                        if (!coversAll(index, subsubClause,indexInfo)) {
-                            return null;
-                        }
-                    }
-                } else {
-                    if (!coversAll(index, subClause, indexInfo)) {
-                        return null;
-                    }
-                }
-            }
-            covered.add(condition);
-            return condition;
-        }
-        assert condition instanceof And;
-        final And<JanusGraphElement> subCondition = new And<>(condition.numChildren());
-        for (final Condition<JanusGraphElement> subClause : condition.getChildren()) {
-            if (coversAll(index,subClause,indexInfo)) {
-                subCondition.add(subClause);
-                covered.add(subClause);
-            }
-        }
-        return subCondition.isEmpty()?null:subCondition;
-    }
-
-    private static boolean coversAll(final MixedIndexType index, Condition<JanusGraphElement> condition,
-                                     IndexSerializer indexInfo) {
-        if (condition.getType()!=Condition.Type.LITERAL) {
-            return StreamSupport.stream(condition.getChildren().spliterator(), false)
-                .allMatch(child -> coversAll(index, child, indexInfo));
-        }
-        if (!(condition instanceof PredicateCondition)) {
-            return false;
-        }
-        final PredicateCondition<RelationType, JanusGraphElement> atom = (PredicateCondition) condition;
-        if (atom.getValue() == null) {
-            return false;
-        }
-
-        Preconditions.checkArgument(atom.getKey().isPropertyKey());
-        final PropertyKey key = (PropertyKey) atom.getKey();
-        final ParameterIndexField[] fields = index.getFieldKeys();
-        final ParameterIndexField match = Arrays.stream(fields)
-            .filter(field -> field.getStatus() == SchemaStatus.ENABLED)
-            .filter(field -> field.getFieldKey().equals(key))
-            .findAny().orElse(null);
-        return match != null && indexInfo.supports(index, match, atom.getPredicate());
-    }
-
-    private JointIndexQuery selectIndices(final Set<IndexType> indexCandidates, final MultiCondition<JanusGraphElement> conditions,
-                                          final Set<Condition> coveredClauses, final MutableBoolean isSorted) {
-        /* TODO: smarter optimization:
-        - use in-memory histograms to estimate selectivity of PredicateConditions and filter out low-selectivity ones
-                if they would result in an individual index call (better to filter afterwards in memory)
-        */
-        if (indexCandidates.size() <= tx.getGraph().getConfiguration().getIndexSelectThreshold()) {
-            return selectIndicesByBruteForce(indexCandidates, conditions, coveredClauses, isSorted);
-        } else {
-            return selectIndicesByApprox(indexCandidates, conditions, coveredClauses, isSorted);
-        }
-    }
-
-    /**
-     * Determine the best jointIndexQuery by enumerating all possibilities with exponential time complexity.
-     * Similar to Weighted Set Cover problem, to find the best choice is NP-Complete, so we should be careful
-     * that the problem size MUST be small, otherwise it is more recommended to use an approximation algorithm.
-     *
-     * @param rawCandidates
-     * @param conditions
-     * @param coveredClauses
-     * @param isSorted
-     * @return
-     */
-    private JointIndexQuery selectIndicesByBruteForce(final Set<IndexType> rawCandidates, final MultiCondition<JanusGraphElement> conditions,
-                                                      final Set<Condition> coveredClauses, final MutableBoolean isSorted) {
-        final JointIndexQuery jointQuery = new JointIndexQuery();
-        final Set<IndexCandidate> indexCandidates = new HashSet<>();
-
-        // validate, enrich index candidates and calculate scores
-        for (final IndexType index : rawCandidates) {
-            try {
-                IndexCandidate ic = new IndexCandidate(index, conditions, serializer);
-                ic.calculateScoreBruteForce();
-                indexCandidates.add(ic);
-            } catch (IllegalArgumentException ignored) {
-                // ignore invalid index candidates
-            }
-        }
-
-        IndexCandidateGroup bestGroup = null;
-        for (Set<IndexCandidate> subset : new PowerSet<IndexCandidate>(indexCandidates)) {
-            if (subset.isEmpty()) continue;
-            final IndexCandidateGroup group = new IndexCandidateGroup(subset);
-            if (group.compareTo(bestGroup) > 0) {
-                bestGroup = group;
-            }
-        }
-
-        if (bestGroup != null) {
-            coveredClauses.addAll(bestGroup.getCoveredClauses());
-            List<IndexCandidate> bestIndexes = new ArrayList<>(bestGroup.getIndexCandidates());
-            // sort indexes by score descending order
-            bestIndexes.sort((a, b) -> Double.compare(b.getScore(), a.getScore()));
-            // isSorted depends on the first index subquery
-            isSorted.setValue(orders.isEmpty() || bestIndexes.get(0).getIndex().isMixedIndex()
-                && indexCoversOrder((MixedIndexType) bestIndexes.get(0).getIndex(), orders));
-            for (IndexCandidate c : bestIndexes) {
-                addToJointQuery(c, jointQuery);
-            }
-        }
-
-        return jointQuery;
-    }
-
-    /**
-     * Iterate over all potential indexes and compute a score based on how many clauses
-     * this index covers. The index with the highest score (as long as it covers at least one additional clause)
-     * is picked and added to the joint query for as long as such exist.
-     * @param rawCandidates
-     * @param conditions
-     * @param coveredClauses
-     * @param isSorted
-     * @return
-     */
-    private JointIndexQuery selectIndicesByApprox(final Set<IndexType> rawCandidates, final MultiCondition<JanusGraphElement> conditions,
-                                                  final Set<Condition> coveredClauses, final MutableBoolean isSorted) {
-        final JointIndexQuery jointQuery = new JointIndexQuery();
-        while (true) {
-            IndexCandidate bestCandidate = null;
-            boolean candidateSupportsSort = false;
-
-            for (final IndexType index : rawCandidates) {
-                try {
-                    final IndexCandidate indexCandidate = new IndexCandidate(index, conditions, serializer);
-
-                    boolean supportsSort = orders.isEmpty() ||
-                        coveredClauses.isEmpty() && index.isMixedIndex() && indexCoversOrder((MixedIndexType) index, orders);
-                    indexCandidate.calculateScoreApproximation(coveredClauses, supportsSort);
-
-                    if (!indexCandidate.isCoveredBy(coveredClauses) && (bestCandidate == null || indexCandidate.getScore() > bestCandidate.getScore())) {
-                        bestCandidate = indexCandidate;
-                        candidateSupportsSort = supportsSort;
-                    }
-                } catch (IllegalArgumentException ignored) {
-                    // ignore invalid index candidates
-                }
-            }
-            if (bestCandidate!=null) {
-                if (coveredClauses.isEmpty()) isSorted.setValue(candidateSupportsSort);
-                coveredClauses.addAll(bestCandidate.getSubCover());
-                addToJointQuery(bestCandidate, jointQuery);
-            } else {
-                break;
-            }
-        }
-        return jointQuery;
-    }
-
-    private void addToJointQuery(final IndexCandidate indexCandidate, final JointIndexQuery jointQuery) {
-        if (indexCandidate.getIndex().isCompositeIndex()) {
-            jointQuery.add((CompositeIndexType) indexCandidate.getIndex(), serializer.getQuery(
-                (CompositeIndexType) indexCandidate.getIndex(), (List<Object[]>) indexCandidate.getSubCondition()));
-        } else {
-            jointQuery.add((MixedIndexType) indexCandidate.getIndex(), serializer.getQuery(
-                (MixedIndexType) indexCandidate.getIndex(), (Condition) indexCandidate.getSubCondition(), orders));
-        }
     }
 }
