+++ janusgraph-core/src/main/java/org/janusgraph/graphdb/query/index/IndexSelectionUtil.java
@@ -1,0 +1,28 @@
+// Copyright 2020 JanusGraph Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package org.janusgraph.graphdb.query.index;
+
+import org.janusgraph.graphdb.internal.OrderList;
+import org.janusgraph.graphdb.types.MixedIndexType;
+
+public class IndexSelectionUtil {
+    public static boolean indexCoversOrder(MixedIndexType index, OrderList orders) {
+        for (int i = 0; i < orders.size(); i++) {
+            if (!index.indexesKey(orders.getKey(i)))
+                return false;
+        }
+        return true;
+    }
+}
