+++ janusgraph-test/src/test/java/org/janusgraph/graphdb/transaction/StandardJanusGraphTxTest.java
@@ -20,6 +20,8 @@ import static org.easymock.EasyMock.replay;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 import org.easymock.EasyMockSupport;
+import org.janusgraph.graphdb.query.index.IndexSelectionStrategy;
+import org.janusgraph.graphdb.query.index.ThresholdBasedIndexSelectionStrategy;
 import org.junit.jupiter.api.Test;
 
 import org.janusgraph.core.RelationType;
@@ -64,6 +66,7 @@ public class StandardJanusGraphTxTest extends EasyMockSupport {
         IDManager idManager = createMock(IDManager.class);
         PropertyKey propertyKey = createMock(PropertyKey.class);
         DefaultSchemaMaker defaultSchemaMaker = createMock(DefaultSchemaMaker.class);
+        IndexSelectionStrategy indexSelectionStrategy = createMock(ThresholdBasedIndexSelectionStrategy.class);
 
         expect(mockGraph.getConfiguration()).andReturn(gdbConfig);
         expect(mockGraph.isOpen()).andReturn(true).anyTimes();
@@ -71,6 +74,7 @@ public class StandardJanusGraphTxTest extends EasyMockSupport {
         expect(mockGraph.getEdgeSerializer()).andReturn(mockEdgeSerializer);
         expect(mockGraph.getIndexSerializer()).andReturn(mockIndexSerializer);
         expect(mockGraph.getIDManager()).andReturn(idManager);
+        expect(mockGraph.getIndexSelector()).andReturn(indexSelectionStrategy);
 
         expect(gdbConfig.getTimestampProvider()).andReturn(tsProvider);
